+++ somutils/testutils.py
@@ -1,8 +1,13 @@
 # -*- coding: utf-8 -*-
 
 import unittest
+from contextlib import contextmanager
 
 from yamlns.testutils import assertNsEqual
+try:
+    from pathlib import Path
+except ImportError:
+    from pathlib2 import Path # Py2
 
 # Readable verbose testcase listing
 unittest.TestCase.__str__ = unittest.TestCase.id
@@ -23,6 +28,33 @@ def _inProduction():
 def destructiveTest(decorated):
     return unittest.skipIf(_inProduction(),
         "Destructive test being run in a production setup!!")(decorated)
+
+@contextmanager
+def temp_path():
+    """
+    Context manager that creates a temporary dir and ensures
+    that all the content is removed after the with scope
+    Returns the pathlib Path of the created dir
+    >>> with temp_path() as tmp:
+    ...     mypath = tmp / 'myfile'
+    ...     nbytes = mypath.write_text('hello world', encoding='utf8')
+    ...     mypath.read_text(encoding='utf8') # returns "hello world"
+    ...     assert mypath.exists(), "Should exists at this point"
+    'hello world'
+
+    >>> assert not mypath.exists(), "Sould not exist at this point"
+    """
+
+    import shutil
+    import tempfile
+
+    path = Path(tempfile.mkdtemp())
+    shutil.rmtree(path, ignore_errors=True)
+    path.mkdir(parents=True, exist_ok=True)
+    try:
+        yield path
+    finally:
+        shutil.rmtree(path, ignore_errors=True)